requirements.txt
streamlit==1.45.1
transformers==4.37.2
torch==2.2.2
numpy<2.0
pandas==2.2.3
pymongo==4.6.3
python-dotenv==1.1.0
